Add files via upload
Important differences from the regular Lion module:
Doubled the movement speed of the Lion sprite.
Reduced the size of the Lion sprite to 75%
Added "goodSpeed" and "badSpeed" variables to randomly generate how fast items will be falling. These are not module specific and can be reused across all characters.
Changed the required amount of Tacos to be collected from 3 to 5.